requirements.txt
Django~=5.0.4
djangorestframework~=3.15.1
asgiref~=3.8.1
django-cors-headers~=4.3.1
django-mysql~=4.12.0
mysqlclient~=2.2.4
sqlparse~=0.5.0
tzdata~=2024.1
auth0-python~=4.7.1
requests~=2.31.0